Determine the project root using __init__.py files
Instead of looking for an assortment of files that usually are located in the
project root of a python project, we instead look for the first parent directory
that does *not* contain an __init__.py file.

This idea was formulated by jamescasbon on
https://github.com/emacs-lsp/lsp-python/issues/5.  I have been using it for the
last couple of months without any issues.  Apart from the simpler logic for
finding the project root this also has the added advantage of correctly
determining the project root for single file python scripts.